requirements.txt
alembic==1.7.7
anyio==3.5.0
asgiref==3.5.0
bcrypt==3.2.0
cffi==1.15.0
click==8.1.3
cryptography==37.0.1
dnspython==2.2.1
ecdsa==0.17.0
email-validator==1.2.0
fastapi==0.75.2
greenlet==1.1.2
h11==0.13.0
httptools==0.4.0
idna==3.3
Mako==1.2.0
MarkupSafe==2.1.1
passlib==1.7.4
psycopg2-binary==2.9.3
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.0
PyJWT==2.3.0
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.36
starlette==0.17.1
typing_extensions==4.2.0
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.2
websockets==10.3
